Add support for getting the underlying listener
Implementing UTransport based on ComparableListener may require
exposing the (interal) ComparableListener type to clients of UTransport
when it comes to invoking the listener.

Added an into_inner function for getting a clone of the reference to
the wrapped listener. This way, client code can easily transform a
ComparableListener back to its underlying listener type.