[#100] Remove support for CloudEvents
The functionality implemented by the cloudevents module did not serve
any real purpose since the introduction of UMessage as the primary
means to pass messages around. UTransport implementations may
or may not choose to map UMessages to CloudEvents before writing to
the transport PDU. The mapping is part of the uProtocol Specification
but the cloudvents module did not specifically support it.

Therefore, CloudEvent support has been removed from up-rust.

Fixes #100