Use Gherkin to define TCK for UUID and UUri serialization
The up-tck component had been using Gherkin for defining the TCK's
test scenarios. The idea was that these could be used for verifying a
language library's compliance with the specification.

This is a first step in the direction described in
https://github.com/eclipse-uprotocol/up-tck/issues/116.

The TCK's test scenarios have been adapted to cover more cases and, in
particular, also cover failure cases. The corresponding unit tests have
been replaced by Cucumber based tests and the Gherkin files have been
amended with OpenFastTrace specification identifiers and are now also
included when doing the requirements tracing run.