[#269] Restrict number of feature permutations to test
Most of the possible permutations of features that are currently
being tested in the nightly job are not relevant in practice.

The set of feature combinations has therefore been drastically reduced
in order to prevent useless combinations from being tested and the
GitHub Agent runner to run out of disk space.